[refactor] : Bottle의 Id 타입 Long 으로 변경

+++ feat/bottle/src/main/kotlin/com/team/bottles/feat/bottle/bottlebox/BottleBoxRoute.kt
@@ -10,7 +10,7 @@ import com.team.bottles.feat.bottle.bottlebox.mvi.BottleBoxSideEffect
 @Composable
 internal fun BottleBoxRoute(
     viewModel: BottleBoxViewModel = hiltViewModel(),
-    navigateToBottle: (Int) -> Unit
+    navigateToBottle: (Long) -> Unit
 ) {
     val uiState by viewModel.state.collectAsStateWithLifecycle()
